refactor to be a more broader component
