+++ src/components/AllRead.tsx
@@ -1,6 +1,6 @@
 import { type FC, useMemo } from 'react';
 import { Constants } from '../utils/constants';
-import { Emoji } from './icons/Emoji';
+import { EmojiText } from './EmojiText';
 
 export const AllRead: FC = () => {
   const emoji = useMemo(
@@ -14,7 +14,7 @@ export const AllRead: FC = () => {
   return (
     <div className="flex flex-1 flex-col items-center justify-center bg-white p-4 text-black dark:bg-gray-dark dark:text-white">
       <div className="mt-2 mb-5 text-5xl">
-        <Emoji emoji={emoji} />
+        <EmojiText text={emoji} />
       </div>
 
       <div className="mb-2 text-xl font-semibold">No new notifications.</div>
