+++ src/components/EmojiText.test.tsx
@@ -1,6 +1,6 @@
 import { render } from '@testing-library/react';
-import { mockDirectoryPath } from '../../__mocks__/utils';
-import { Emoji, type IEmoji } from './Emoji';
+import { mockDirectoryPath } from '../__mocks__/utils';
+import { EmojiText, type IEmojiText } from './EmojiText';
 
 describe('components/icons/Emoji.tsx', () => {
   beforeEach(() => {
@@ -8,10 +8,10 @@ describe('components/icons/Emoji.tsx', () => {
   });
 
   it('should render', () => {
-    const props: IEmoji = {
-      emoji: '🍺',
+    const props: IEmojiText = {
+      text: '🍺',
     };
-    const tree = render(<Emoji {...props} />);
+    const tree = render(<EmojiText {...props} />);
     expect(tree).toMatchSnapshot();
   });
 });
